+++ inc/uniform.h
@@ -2,14 +2,14 @@
 #define UNIFORM_H
 
 // Sample uniformly on [0,1)
-double rand_unif_half_open(void);
+float rand_unif_half_open(void);
 
 // Sample uniformly on [0,1]
-double rand_unif_closed(void);
+float rand_unif_closed(void);
 
 // Sample uniformly on (0,1)
-double rand_unif_open(void);
+float rand_unif_open(void);
 
-int uniform_multiple(double* arr, int N);
+int uniform_multiple(float* arr, int N);
 
 #endif
